pyproject.toml
[tool.poetry]
name = "frequency"
version = "0.1.0"
description = "Effiently serve LoRA tuned models"
authors = ["Patrick Barker <patrickbarkerco@gmail.com>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.108.0"
sqlalchemy = "^2.0.25"
uvicorn = "^0.25.0"

[tool.poetry.group.gcp.dependencies]
google-cloud-storage = "^2.14.0"


[tool.poetry.group.hf.dependencies]
transformers = "^4.36.2"
peft = "^0.7.1"


[tool.poetry.group.runpod.dependencies]
runpod = "^1.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
